partition manager and make vfs read offset 64 bit

+++ include/kernel/fs/vfs.h
@@ -12,8 +12,8 @@ struct dirent
 
 struct fs_node;
 
-typedef uint32_t (*read_type_t)(struct fs_node*,uint32_t,uint32_t,uint8_t*);
-typedef uint32_t (*write_type_t)(struct fs_node*,uint32_t,uint32_t,uint8_t*);
+typedef uint64_t (*read_type_t)(struct fs_node*,uint64_t,uint64_t,uint8_t*);
+typedef uint64_t (*write_type_t)(struct fs_node*,uint64_t,uint64_t,uint8_t*);
 typedef void (*open_type_t)(struct fs_node*);
 typedef void (*close_type_t)(struct fs_node*);
 typedef void (*create_file_type_t)(struct fs_node*, char* name);
@@ -57,8 +57,8 @@ extern fs_node_t *fs_root; // The root of the filesystem.
 // Standard read/write/open/close functions. Note that these are all suffixed with
 // _fs to distinguish them from the read/write/open/close which deal with file descriptors
 // not file nodes.
-uint32_t read_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
-uint32_t write_fs(fs_node_t *node, uint32_t offset, uint32_t size, uint8_t *buffer);
+uint64_t read_fs(fs_node_t *node, uint64_t offset, uint64_t size, uint8_t *buffer);
+uint64_t write_fs(fs_node_t *node, uint64_t offset, uint64_t size, uint8_t *buffer);
 void open_fs(fs_node_t *node, uint8_t read, uint8_t write);
 void close_fs(fs_node_t *node);
 void truncate_fs(fs_node_t* node, int length);
